requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-chess==1.999
